Split shared/testing into separate client/server modules.
This also gets rid of a conditional import.
