+++ shell/server/main.ts
@@ -107,4 +107,5 @@ import "../imports/server/transfers-server.js";
 import "../imports/shared/admin.js";
 import "../imports/shared/dev-accounts.js";
 import "../imports/shared/grain-shared.js";
-import "../imports/shared/testing.js";
+
+import "../imports/server/testing.js";
